feat: refactor hero animation, divide hooks
