+++ src/components/pages/home/hero/animation/animation.jsx
@@ -1,43 +1,14 @@
-import { useGSAP } from '@gsap/react';
-import { Layout, Fit, Alignment, EventType } from '@rive-app/canvas';
-import { useStateMachineInput } from '@rive-app/react-canvas';
 import { clsx } from 'clsx';
 import { StaticImage } from 'gatsby-plugin-image';
-import gsap from 'gsap';
-import ScrollTrigger from 'gsap/ScrollTrigger';
-import React, { useRef, useCallback, useState, useEffect } from 'react';
+import React, { useRef } from 'react';
 
 import Button from 'components/shared/button';
 import RiveAnimation from 'components/shared/rive-animation';
-import useRiveAnimation from 'hooks/use-rive-animation';
 
 import { firstSectionData, secondSectionData, thirdSectionData } from './data';
-
-const useAnimation = (artboard) => {
-  const { riveInstance, wrapperRef, animationRef, setRiveInstance } = useRiveAnimation({});
-
-  const riveAnimationProps = {
-    src: '/animations/pages/home/hero/new-hero.riv',
-    artboard,
-    autoplay: false,
-    stateMachines: 'SM',
-    layout: new Layout({
-      fit: Fit.Fill,
-      alignment: Alignment.Center,
-    }),
-    onLoad: () => {
-      riveInstance?.resizeDrawingSurfaceToCanvas();
-    },
-  };
-
-  return {
-    riveInstance,
-    wrapperRef,
-    animationRef,
-    setRiveInstance,
-    riveAnimationProps,
-  };
-};
+import useAnimation from './use-animation';
+import useAnimationSetup from './use-animation-setup';
+import useGSAPAnimations from './use-gsap-animations';
 
 const Animation = () => {
   // define refs for all animation elements
@@ -59,6 +30,20 @@ const Animation = () => {
 
   const lightImageRef = useRef(null);
 
+  const refs = {
+    animationRef,
+    containerRef,
+    cardBlueRef,
+    cardPurpleRef,
+    cardCodeRef,
+    phoneRef,
+    inboxRef,
+    developersRef,
+    productTeamsRef,
+    endUsersRef,
+    lightImageRef,
+    snapRef,
+  };
   // initialize animations
   const {
     riveInstance: cardBlueAnimationInstance,
@@ -100,687 +85,23 @@ const Animation = () => {
     riveAnimationProps: inboxAnimationProps,
   } = useAnimation('inbox');
 
-  // define ready states for animations
-  const [isCodeAnimationReady, setIsCodeAnimationReady] = useState(false);
-  const [isBlueAnimationReady, setIsBlueAnimationReady] = useState(false);
-  const [isPurpleAnimationReady, setIsPurpleAnimationReady] = useState(false);
+  const { isCodeAnimationReady, isBlueAnimationReady, isPurpleAnimationReady, stateMachineInputs } =
+    useAnimationSetup({
+      cardBlueAnimationInstance,
+      cardPurpleAnimationInstance,
+      cardCodeAnimationInstance,
+      phoneAnimationInstance,
+      inboxAnimationInstance,
+    });
 
-  useEffect(() => {
-    if (cardBlueAnimationInstance?.readyForPlaying) {
-      setIsBlueAnimationReady(true);
-    }
-  }, [cardBlueAnimationInstance]);
-
-  useEffect(() => {
-    if (cardPurpleAnimationInstance?.readyForPlaying) {
-      setIsPurpleAnimationReady(true);
-    }
-  }, [cardPurpleAnimationInstance]);
-
-  useEffect(() => {
-    if (cardCodeAnimationInstance?.readyForPlaying) {
-      setIsCodeAnimationReady(true);
-    }
-  }, [cardCodeAnimationInstance]);
-
-  // create stateMachineInputs for all animations
-  const cardPurpleFloating = useStateMachineInput(
-    cardPurpleAnimationInstance,
-    'SM',
-    'floating',
-    true
-  );
-  const cardPurpleSparkle = useStateMachineInput(
-    cardPurpleAnimationInstance,
-    'SM',
-    'sparkle',
-    true
-  );
-  const cardBlueFloating = useStateMachineInput(cardBlueAnimationInstance, 'SM', 'floating', true);
-  const cardBlueDisabled = useStateMachineInput(cardBlueAnimationInstance, 'SM', 'mobile', true);
-  const cardCodeFloating = useStateMachineInput(cardCodeAnimationInstance, 'SM', 'floating', true);
-
-  const cardCodeChange = useStateMachineInput(cardCodeAnimationInstance, 'SM', 'code-change');
-  const cardPurpleChangeNumbers = useStateMachineInput(
-    cardPurpleAnimationInstance,
-    'SM',
-    'numbers'
-  );
-
-  const cardBlueChangeName = useStateMachineInput(cardBlueAnimationInstance, 'SM', 'name-change');
-  const cardPurpleChangeName = useStateMachineInput(cardPurpleAnimationInstance, 'SM', 'name');
-  const cardPurpleChangeTag = useStateMachineInput(cardPurpleAnimationInstance, 'SM', 'tag', true);
-  const cardPurpleChangeAvatar = useStateMachineInput(
-    cardPurpleAnimationInstance,
-    'SM',
-    'avatar',
-    false
-  );
-  const cardPurpleChangeDescription = useStateMachineInput(
-    cardPurpleAnimationInstance,
-    'SM',
-    'description',
-    true
-  );
-
-  const phoneNotification = useStateMachineInput(
-    phoneAnimationInstance,
-    'SM',
-    'notification',
-    true
-  );
-  const inboxNotification = useStateMachineInput(
-    inboxAnimationInstance,
-    'SM',
-    'notification',
-    true
-  );
-
-  const phoneReset = useStateMachineInput(phoneAnimationInstance, 'SM', 'reset', true);
-  const inboxReset = useStateMachineInput(inboxAnimationInstance, 'SM', 'reset', true);
-
-  // create listeners for all animations
-  const cardCodeAnimationListener = useCallback(
-    (event) => {
-      if (event.data.name === 'numbers' && cardPurpleChangeNumbers) {
-        cardPurpleChangeNumbers.fire();
-      }
-    },
-    [cardPurpleChangeNumbers]
-  );
-
-  const cardBlueAnimationListener = useCallback(
-    (event) => {
-      if (event.data.name === 'name' && cardPurpleChangeName) {
-        cardPurpleChangeName.fire();
-      }
-      if (event.data.name === 'tag' && cardPurpleChangeTag) {
-        cardPurpleChangeTag.value = !cardPurpleChangeTag.value;
-      }
-      if (event.data.name === 'avatar' && cardPurpleChangeAvatar) {
-        cardPurpleChangeAvatar.value = !cardPurpleChangeAvatar.value;
-      }
-      if (event.data.name === 'description' && cardPurpleChangeDescription) {
-        cardPurpleChangeDescription.value = !cardPurpleChangeDescription.value;
-      }
-    },
-    [cardPurpleChangeName, cardPurpleChangeTag, cardPurpleChangeAvatar, cardPurpleChangeDescription]
-  );
-
-  if (cardCodeAnimationInstance) {
-    cardCodeAnimationInstance.on(EventType.RiveEvent, cardCodeAnimationListener);
-  }
-
-  if (cardBlueAnimationInstance) {
-    cardBlueAnimationInstance.on(EventType.RiveEvent, cardBlueAnimationListener);
-  }
-
-  gsap.registerPlugin(useGSAP, ScrollTrigger);
-
-  // GSAP section for animation triggers
-  useGSAP(
-    () => {
-      const height = animationRef.current.offsetHeight;
-
-      const offsetHeight = containerRef.current.offsetTop;
-
-      const isAnimationPlayable = window.innerWidth >= 1024;
-
-      if (isAnimationPlayable) {
-        ScrollTrigger.create({
-          trigger: containerRef.current,
-          start: `top ${offsetHeight}px`,
-          end: `+=${offsetHeight}px`,
-          onEnter: () => {
-            if (cardPurpleFloating) {
-              cardPurpleFloating.value = false;
-            }
-            if (cardBlueFloating) {
-              cardBlueFloating.value = false;
-            }
-            if (cardCodeFloating) {
-              cardCodeFloating.value = false;
-            }
-          },
-          onLeaveBack: () => {
-            if (cardPurpleFloating) {
-              cardPurpleFloating.value = true;
-            }
-            if (cardBlueFloating) {
-              cardBlueFloating.value = true;
-            }
-            if (cardCodeFloating) {
-              cardCodeFloating.value = true;
-            }
-          },
-        });
-
-        ScrollTrigger.create({
-          trigger: containerRef.current,
-          start: `${height * 0.99}px`,
-          end: `+=${height * 0.1}px`,
-          onEnter: () => {
-            let interval = null;
-
-            if (cardCodeChange && animationInterval.current === null) {
-              let counter = 0;
-
-              cardCodeChange.fire();
-
-              interval = setInterval(() => {
-                cardCodeChange.fire();
-                counter += 1;
-                if (counter >= 2) {
-                  clearInterval(interval);
-                }
-              }, 3000);
-
-              animationInterval.current = interval;
-            }
-
-            return () => {
-              clearInterval(interval);
-              animationInterval.current = null;
-            };
-          },
-          onLeave: () => {
-            clearInterval(animationInterval.current);
-            animationInterval.current = null;
-
-            if (cardPurpleSparkle) {
-              cardPurpleSparkle.value = false;
-            }
-            if (cardBlueDisabled) {
-              cardBlueDisabled.value = false;
-            }
-          },
-          onLeaveBack: () => {
-            if (cardPurpleSparkle) {
-              cardPurpleSparkle.value = true;
-            }
-
-            if (cardBlueDisabled) {
-              cardBlueDisabled.value = true;
-            }
-          },
-        });
-
-        ScrollTrigger.create({
-          trigger: containerRef.current,
-          start: `${height * 1.99}px`,
-          end: `+=${height * 0.1}px`,
-          onEnter: () => {
-            let interval = null;
-
-            if (cardBlueChangeName && animationInterval.current === null) {
-              let counter = 0;
-
-              cardBlueChangeName.fire();
-
-              interval = setInterval(() => {
-                cardBlueChangeName.fire();
-                counter += 1;
-                if (counter >= 2) {
-                  clearInterval(interval);
-                }
-              }, 3000);
-
-              animationInterval.current = interval;
-            }
-
-            return () => {
-              clearInterval(interval);
-              animationInterval.current = null;
-            };
-          },
-          onLeave: () => {
-            clearInterval(animationInterval.current);
-            animationInterval.current = null;
-
-            if (cardBlueDisabled) {
-              cardBlueDisabled.value = true;
-            }
-          },
-          onLeaveBack: () => {
-            if (phoneReset) {
-              phoneReset.fire();
-            }
-            if (inboxReset) {
-              inboxReset.fire();
-            }
-            if (cardBlueDisabled) {
-              cardBlueDisabled.value = false;
-            }
-          },
-        });
-
-        ScrollTrigger.create({
-          trigger: containerRef.current,
-          start: `${height * 2.99}px`,
-          end: `+=${height * 0.1}px`,
-          onEnter: () => {
-            if (phoneNotification) {
-              phoneNotification.fire();
-            }
-            if (inboxNotification) {
-              inboxNotification.fire();
-            }
-            if (cardPurpleSparkle) {
-              cardPurpleSparkle.value = true;
-            }
-          },
-          onLeaveBack: () => {
-            if (cardPurpleSparkle) {
-              cardPurpleSparkle.value = false;
-            }
-          },
-          onEnterBack: () => {
-            if (cardBlueDisabled) {
-              cardBlueDisabled.value = false;
-            }
-          },
-        });
-      }
-    },
-    {
-      dependencies: [
-        cardPurpleFloating,
-        cardPurpleSparkle,
-        cardBlueFloating,
-        cardCodeFloating,
-        cardBlueChangeName,
-        phoneNotification,
-        inboxNotification,
-        animationInterval,
-        phoneReset,
-        inboxReset,
-        cardBlueDisabled,
-      ],
-      scope: containerRef.current,
-    }
-  );
-
-  useGSAP(
-    () => {
-      if (isCodeAnimationReady && isBlueAnimationReady && isPurpleAnimationReady) {
-        gsap.to(lightImageRef.current, {
-          opacity: 1,
-          duration: 0.3,
-          ease: 'power2.inOut',
-        });
-      }
-    },
-    {
-      dependencies: [isCodeAnimationReady, isBlueAnimationReady, isPurpleAnimationReady],
-      scope: lightImageRef.current,
-    }
-  );
-
-  // GSAP section for animation positioning
-  useGSAP(
-    () => {
-      const height = animationRef.current.offsetHeight;
-
-      const offsetHeight = containerRef.current.offsetTop;
-
-      const isTablet = window.innerWidth < 1279;
-
-      const isAnimationPlayable = window.innerWidth >= 1024;
-
-      const containerHeight = height * 3;
-
-      const snapTo = [
-        0,
-        (offsetHeight + height) / (containerHeight + offsetHeight),
-        (offsetHeight + height * 2) / (containerHeight + offsetHeight),
-        1,
-      ];
-
-      if (isAnimationPlayable) {
-        ScrollTrigger.create({
-          trigger: containerRef.current,
-          start: 'top top',
-          pin: animationRef.current,
-          pinSpacing: false,
-          end: 'bottom bottom',
-        });
-
-        ScrollTrigger.create({
-          trigger: snapRef.current,
-          start: 'top top',
-          end: 'bottom bottom',
-          snap: {
-            snapTo,
-            duration: 0.5,
-            delay: 0,
-            ease: 'power2.inOut',
-          },
-        });
-
-        // first step
-        gsap.to(lightImageRef.current, {
-          immediateRender: false,
-          opacity: 0,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `top ${offsetHeight}px`,
-            end: `+=${offsetHeight * 0.5}px`,
-          },
-        });
-
-        gsap.to(cardPurpleRef.current, {
-          startAt: { yPercent: 0, left: '19.473%', scale: 1 },
-          top: '50%',
-          yPercent: -49,
-          left: '22.708%',
-          skewY: -15,
-          scale: 1.47,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `top ${offsetHeight}px`,
-            end: `+=${offsetHeight}px`,
-          },
-        });
-
-        gsap.to(cardBlueRef.current, {
-          startAt: { yPercent: 0, x: 0, left: '35.713%', scale: 1 },
-          top: '50%',
-          yPercent: -81,
-          left: '29.74%',
-          x: 0,
-          skewY: -15,
-          scale: 1.44,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `top ${offsetHeight}px`,
-            end: `+=${offsetHeight}px`,
-          },
-        });
-
-        gsap.to(cardCodeRef.current, {
-          startAt: { yPercent: 0, right: '23.061%', scale: 1 },
-          top: '50%',
-          yPercent: -81,
-          right: '22.604%',
-          skewY: -15,
-          scale: 1.42,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `top ${offsetHeight}px`,
-            end: `+=${offsetHeight}px`,
-          },
-        });
-
-        // developers step
-        gsap.to(cardPurpleRef.current, {
-          immediateRender: false,
-          left: '34.4%',
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: 'top',
-            end: `+=${height}px`,
-          },
-        });
-        gsap.to(cardCodeRef.current, {
-          immediateRender: false,
-          right: '50.2%',
-          x: 0,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: 'top',
-            end: `+=${height}px`,
-          },
-        });
-        gsap.to(cardBlueRef.current, {
-          immediateRender: false,
-          left: '22.172%',
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: 'top',
-            end: `+=${height * 0.5}px`,
-          },
-        });
-        gsap.to(cardBlueRef.current, {
-          immediateRender: false,
-          opacity: 0,
-          x: 0,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 0.35}px`,
-            end: `+=${height * 0.15}px`,
-          },
-        });
-        gsap.to(developersRef.current, {
-          startAt: { x: 100, opacity: 0, pointerEvents: 'none' },
-          x: 0,
-          opacity: 1,
-          pointerEvents: 'auto',
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 0.5}px`,
-            end: `+=${height * 0.5}px`,
-          },
-        });
-
-        // product teams step
-        gsap.to(developersRef.current, {
-          immediateRender: false,
-          x: -50,
-          opacity: 0,
-          pointerEvents: 'none',
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height}px`,
-            end: `+=${height * 0.2}px`,
-          },
-        });
-        gsap.to(cardCodeRef.current, {
-          immediateRender: false,
-          x: -50,
-          opacity: 0,
-          pointerEvents: 'none',
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height}px`,
-            end: `+=${height * 0.2}px`,
-          },
-        });
-        gsap.to(cardBlueRef.current, {
-          startAt: { left: '43.2%', x: 140 },
-          immediateRender: false,
-          x: 0,
-          left: isTablet ? '50.2%' : '43.2%',
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 1.2}px`,
-            end: `+=${height * 0.8}px`,
-          },
-        });
-        gsap.to(cardBlueRef.current, {
-          startAt: { opacity: 0 },
-          immediateRender: false,
-          opacity: 1,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 1.2}px`,
-            end: `+=${height * 0.2}px`,
-          },
-        });
-        gsap.to(cardPurpleRef.current, {
-          immediateRender: false,
-          left: isTablet ? '43.3%' : '36.3%',
-          pointerEvents: 'none',
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height}px`,
-            end: `+=${height}px`,
-          },
-        });
-        gsap.to(productTeamsRef.current, {
-          startAt: { x: 140, opacity: 0, pointerEvents: 'none' },
-          x: 0,
-          opacity: 1,
-          pointerEvents: 'auto',
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 1.2}px`,
-            end: `+=${height * 0.8}px`,
-          },
-        });
-
-        // end users step
-        gsap.to(productTeamsRef.current, {
-          immediateRender: false,
-          y: -60,
-          opacity: 0,
-          pointerEvents: 'none',
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.0}px`,
-            end: `+=${height * 0.2}px`,
-          },
-        });
-        gsap.to(cardBlueRef.current, {
-          immediateRender: false,
-          y: -700,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.0}px`,
-            end: `+=${height * 0.5}px`,
-          },
-        });
-        gsap.to(cardBlueRef.current, {
-          immediateRender: false,
-          opacity: 0,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.0}px`,
-            end: `+=${height * 0.2}px`,
-          },
-        });
-        gsap.to(endUsersRef.current, {
-          startAt: { top: '0.926%', pointerEvents: 'none' },
-          top: '11.111%',
-          pointerEvents: 'auto',
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.4}px`,
-            end: `+=${height * 0.6}px`,
-          },
-        });
-        gsap.to(endUsersRef.current, {
-          startAt: { opacity: 0 },
-          opacity: 1,
-          ease: 'none',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.4}px`,
-            end: `+=${height * 0.2}px`,
-          },
-        });
-        gsap.to(cardPurpleRef.current, {
-          immediateRender: false,
-          left: '36.3%',
-          yPercent: -35.3,
-          pointerEvents: 'auto',
-          scale: 1.58,
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.3}px`,
-            end: `+=${height * 0.7}px`,
-          },
-        });
-
-        gsap.to(phoneRef.current, {
-          startAt: { x: 150, opacity: 0, pointerEvents: 'none' },
-          x: 0,
-          opacity: 1,
-          pointerEvents: 'auto',
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.3}px`,
-            end: `+=${height * 0.7}px`,
-          },
-        });
-
-        gsap.to(inboxRef.current, {
-          startAt: { x: -150, opacity: 0, pointerEvents: 'none' },
-          x: 0,
-          opacity: 1,
-          pointerEvents: 'auto',
-          ease: 'none',
-          transformOrigin: 'center center',
-          scrollTrigger: {
-            trigger: containerRef.current,
-            scrub: true,
-            start: `${height * 2.3}px`,
-            end: `+=${height * 0.7}px`,
-          },
-        });
-      }
-    },
-    {
-      scope: containerRef.current,
-    }
-  );
+  useGSAPAnimations({
+    refs,
+    animationInterval,
+    stateMachineInputs,
+    isCodeAnimationReady,
+    isBlueAnimationReady,
+    isPurpleAnimationReady,
+  });
 
   return (
     <>
